chore: Dynamic -> Generated
